requirements.txt
Pillow
Django==2.2.*
djangorestframework==3.12.2
django-rest-auth==0.9.5
django-restql==0.11.0
django-filter==2.4.0
django-admin-rangefilter==0.6.3
